First commit
SUPERSHAPE EXPORTER

Supershape Exporter is a custom software to export supershapes in OBJ
format, and it is based on the supershape by Paul Bourke. The main
formula was first proposed by Johan Gielis around 2000, suggesting that
the formula can be used to describe many complex shapes and curves that
are found in nature.

It was inspired by Coding Challenge #26: 3D Supershapes
Find it here: https://youtu.be/akM4wMZIBWg